Add send and validate protocol exchange
This commit adds the first step of the SSH2 protocol. We can now send an
identification string for SSH2 and receive one and validate it.

This file might get pretty big, so it is probably prudent to break off
this connection package into more files for each step in the
authentication process.

We will also start storing state for each connection, such as the MAC
algorithm and any key we need to do kex or ciphers. It would also be
nice to populate it with any (perhaps useful at some stage) data we
receive, so figuring out some sort of shared struct for that would be
ideal.

At this point, an OpenSSH client can connect to our server and get to
the key exchange process!